src/main/java/ru/job4j/cars/service/SimplePostService.java
package ru.job4j.cars.service;

import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;
import ru.job4j.cars.model.Post;
import ru.job4j.cars.repository.PostRepository;

import java.util.List;
import java.util.Optional;

@Service
@AllArgsConstructor
public class SimplePostService implements PostService {

    private final PostRepository postRepository;

    @Override
    public Post save(Post post) {
        return postRepository.save(post);
    }

    @Override
    public List<Post> findAll() {
        return postRepository.findAll();
    }

    @Override
    public Optional<Post> findById(int id) {
        return postRepository.findById(id);
    }

    @Override
    public List<Post> findAllForLastDay() {
        return postRepository.findAllForLastDay();
    }

    @Override
    public List<Post> findAllWithPhoto() {
        return postRepository.findAllWithPhoto();
    }

    @Override
    public List<Post> findByCarMark(String name) {
        return postRepository.findByCarMark(name);
    }

    @Override
    public void changeStatus(int id, boolean status) {
        postRepository.changeStatus(id, status);
    }

}
